pyproject.toml
[project]
name = "yolov1"
version = "0.1.0"
description = "YOLOv1 implementation from scratch in PyTorch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pillow>=9.5.0",
    "numpy>=1.24.0",
    "datasets>=2.14.0",
]

[tool.uv]
package = false
